GH-8 Fix AppVeyor build on Perl 5.12 and older The updates in release 1.86_01 that allow META.yml to be correctly generated when building from repository, broken AppVeyor builds on Strawberry Perl 5.12, 5.10 and 5.8 because of problems installing Module::Install module and its dependencies. We now install a certain version of ExtUtil::MakeMaker and direct cpanm with certain options to get Module::Install installed.
64bit builds with Strawberr Perl 5.10 and 5.8 are now excluded
since those builds are not available.

Module::Install is now required dependency when building from
a repository clone. Directory inc/ was removed from repository
and it will be re-generated by Module::Install.

Because inc/ is no longer part of repository, "." that was
added to lib path in Makefile.PL in version 1.81, was removed.