Prevent memory leak in OCSP_cert2ids() and OCSP_response_verify() (#20)
In SSLeay.xs, the memory used by the X509 object returned by
find_issuer() is allocated on the stack when the issuer exists in the
given chain, and is allocated on the heap when it exists in the store.
Functions that call find_issuer() therefore can't be sure whether to
free this memory after use, leading to a memory leak.

Use X509_dup() to guarantee that the memory used by the return value is
allocated on the heap, then use X509_free() to free that memory wherever
find_issuer() is called (currently in SSL_OCSP_cert2ids() and
SSL_OCSP_response_verify()).

This closes RT#125273. Thanks to Steffen Ullrich for the report and
patch.