Set security level to 1 in tests that use pre-generated RSA keys (#53)
The RSA keys generated for the test suite are 1024 bits long, but some
distributions (notably Debian and RHEL) are now building their OpenSSL
packages with security level 2 as the default, which prohibits the use
of RSA keys under 2048 bits in length and causes tests that load these
keys to fail. Set the security level to 1 (which permits the use of
weaker keys) in tests that use the test suite's keys.

This closes RT#126987. Thanks to Petr Pisar (in RT#126270) and Damyan
Ivanov (in RT#126987) for the reports and patches, and to Damyan Ivanov
for the preferred patch.