Declare variable n_a whenever POPpx macro is used
The SSLeay.xs functions ssleay_set_psk_client_callback_invoke and
ssleay_ctx_set_psk_client_callback_invoke use the POPpx macro, whose API
changed in Perl 5.8.8:

https://www.nntp.perl.org/group/perl.perl5.porters/2006/03/msg110764.html

Before 5.8.8, the POPpx macro required a variable n_a (of type STRLEN)
to be in scope, but no such variable is in scope in these functions.
Define such a variable in these functions (as it is in other functions
in SSLeay.xs) to fix compatibility with Perl 5.8.7 and below.

This fixes RT#128030. Thanks to Graham Ollis for the report.