Always show libssl information in test suite output
The Perl and libssl version information printed near the start of the
test suite is useful (especially for identifying smoke test failures) -
print it as early as possible in the test suite, and don't suppress it
when Test::Exception isn't installed.

Also, output the value of OPENSSL_ENGINES_DIR as part of the libssl
information, if available.

Closes #109.