Support finding OpenSSL libraries using ExtUtils::PkgConfig
If ExtUtils::PkgConfig is available and there is a pkg-config file
for openssl, use the libraries specified rather than assuming which
libraries need to be linked against. This requires ExtUtils::PkgConfig
version 1.16 or later.

This may be useful for example in the case where OpenSSL is built without
compression support and there may be no libz.so to link against.

In the absence of a recent enough version of ExtUtils::PkgConfig, the
Makefile.PL falls back to the existing configuration method, so adding
this enhancement should be harmless.