Replace Twitter chain with our own chain
The Twitter's certiticate expires every year and thus tests fail when
it expires.

This patch replaces it with a chain of certificates that we generate
(and can regenerate with t/data/Makefile anytime) with a 10-year
validity.

It also adjusts other tests to use a different certificate from the
Twitter's one. That allows us to remove it completely.

<https://rt.cpan.org/Ticket/Display.html?id=129201>
<https://github.com/radiator-software/p5-net-ssleay/pull/128>