09_ctx_new.t: Allow getting any valid min/max protocol version
In t/local/09_ctx_new.t, rather than checking that the functions
(CTX_)get_min_proto_version and (CTX_)get_max_proto_version return
0x0000 (indicating the lowest and highest versions supported by libssl
respectively, which is not the case if a run-time configuration is
enforcing a different minimum or maximum), just check whether the
value is one of those mentioned on the SSL_CTX_set_min_proto_version(3)
man page.

This partially fixes RT#128025 and closes #101. Thanks to Slaven Rezić
and Dmytro Zagashev for the downstream reports.