Clarify licensing terms (Artistic License 2.0)
According to Git commit 5a7b40b8 (previously SVN commit r426),
Net-SSLeay was relicensed from the OpenSSL License to the Artistic
License 2.0 in discussions that included Net-SSLeay's original author,
Sampo Kellomäki. Since then, the module distribution's metadata has been
vague or inconsistent about its licensing terms: in some places it is
explicitly stated that the distribution is released under the terms of
the Artistic License 2.0; others imply the Perl 5 License (i.e.
dual-licensed under the Artistic License 1.0 and GPL version 1 or
later), the OpenSSL License (still), or a combination of the three.

This commit clarifies that Net-SSLeay is released under the terms of the
Artistic License 2.0, and that any contributions made to Net-SSLeay are
assumed to be provided under those terms. Copyright notices have also
been standardised across all source files and documentation.

This fixes RT#106314. Thanks to Kent Fredric for pointing out the
ambiguity.