Use 127.0.0.1 instead of localhost in test suite
t/local/06_tcpecho.t and t/local/07_sslecho.t try to connect to the
hostname localhost instead of the IP address 127.0.0.1 in some tests,
causing the implicit or explicit call to gethostbyname() to fail when
the test suite is executed inside a network sandbox (e.g. when
installing Net-SSLeay via Portage on Gentoo).

Replace all remaining instances of localhost with 127.0.0.1, in line
with the other tests in the test suite, which pass in a network sandbox.

This fixes RT#128207. Thanks to Kent Fredric for the downstream report.