Replace HTTPS hosts in external tests
HTTPS servers are no longer online at alioth.debian.org or
banking.wellsfargo.com, so don't attempt to connect to them in the
external tests.

In case the others go offline too, replace them with the following,
which ought not to disappear any time soon:

  * www.google.com
  * www.microsoft.com
  * www.kernel.org

The lists of hosts in t/external/20_cert_chain.t and t/external/ocsp.t
are unchanged, since those tests rely on the properties of those
particular hosts or the certificates they present.

Closes #26.