Test::Net::SSLeay: add tcp_socket helper function
Many scripts in the test suite need to open TCP server and client
sockets to test Net::SSLeay's functionality, but there's inconsistency
in how that is performed:

* Creating the socket: some scripts use Perl's builtin socket functions
  and the Socket module; some use IO::Socket::INET.
* Finding a free port: some scripts define a hard-coded port number,
  some cycle through a range of port numbers looking for a free one. The
  ones that define a hard-coded port number often fail on CPAN Testers'
  infrastructure, as the port is already in use by another copy of the
  test suite.

Add a helper function, tcp_socket, to Test::Net::SSLeay that creates a
socket listening on a randomly-chosen free TCP port, and provides
convenience methods for accepting and creating connections to the server
socket. The functionality is provided by Test::Net::SSLeay::Socket, and
implemented using Perl's builtin socket functions and the Socket module;
no non-core dependencies are introduced.

Closes #172.