.appveyor.yml: don't silently succeed when make fails
In the AppVeyor configuration, "make" is invoked via Perl so that the
correct program can be used depending on the version of Strawberry Perl
in use in the environment (sometimes it's dmake, sometimes it's gmake).
Because the return value of system() is never checked, the Perl command
always succeeds, and AppVeyor silently (and incorrectly) continues
running the job even when the call to make failed.

Explicitly exit from Perl with a return code based on that of the
command invoked by system(), so failures in make are correctly
propagated to AppVeyor's shell and jobs fail when they should.

Closes #176.