Test::Net::SSLeay: add data_file_path helper function
Add a data_file_path helper function to Test::Net::SSLeay, which returns
the relative path to a given file in the test suite data directory
(t/local/). This eliminates the need to directly import File::Spec in
test scripts, so remove any instances of that module being imported too.