Use BN_EGNCB_new() from OpenSSL 1.1.0-pre1 onwards
OpenSSL 1.1.0-pre1 changed the way in which BN_GENCB structures are
created, from allowing them to be declared directly to requiring the
use of BN_EGNCB_new(). The version guard controlling which of these
methods is used in SSLeay.xs is incorrect (1.1.0), resulting in
compile-time errors with OpenSSL versions 1.1.0-pre1 to 1.1.0-pre6.
Correct the version guard to eliminate the compilation failures.

Closes #199.