Don't version-guard SSL_get_SSL_CTX and SSL_ctrl
The OPENSSL_VERSION_NUMBER guard for the SSL_get_SSL_CTX and SSL_ctrl
functions implies they were added in OpenSSL 0.9.8f, but they're
actually present in OpenSSL 0.9.8. Remove the version guard for these
functions, and clarify in the documentation that SSL_set_SSL_CTX is the
only function of the three that requires OpenSSL >= 0.9.8f.

Closes #217.