Makefile.PL: ensure libssl headers exist before writing Makefile
Before writing a Makefile, check that ssleay_get_build_opts() returned
an include directory (inc_path) and display an error message if it did
not. This avoids compilation errors in the situation where the libssl
libraries are present but the headers are not, e.g. on various Linux
distrubutions when the OpenSSL development package is not installed.

Fixes RT#105189.