39_pkcs12.t: validate CA certificate CNs with all libssl versions
When verifying common names in the PKCS#12 certificate chain, account
for the fact that PKCS12_parse() returns the CA certificate chain in a
different order in different versions of libssl.

Since the CA certificate chain tests now cover all versions of OpenSSL
and LibreSSL, don't skip the order check for versions of OpenSSL before
1.0.0, and remove the generic tests that ensure X509_NAME_oneline()
returns a valid-looking common name for each of the CA certificates
(since they're now redundant).