36_verify.t: don't expect legacy verifier flag with LibreSSL 3.4.1
From LibreSSL 3.4.1 onwards, the use of the legacy X.509 verifier is no
longer exposed in the return value of X509_VERIFY_PARAM_get_flags(),
thus reverting to LibreSSL's pre-3.3.2 behaviour. Only expect the
X509_V_FLAG_LEGACY_VERIFY flag to be set in this return value for
versions of LibreSSL between 3.3.2 and 3.4.0.

Closes #324.