Handle vendor OpenSSL 1.1.x on OpenVMS
There is now a product oddly named SSL111 based on OpenSSL 1.1.x
so we need to look in slightly different places to find it.

Also, there is port to x86_64 now in beta, and the executable will
be located accordingly, so add that to the list of places to look.

And ssl.h lives at include/ssl.h, not include/openssl/ssl.h.