Support LibreSSL 3.2 series
Advertise support for LibreSSL 3.2 (with the exception of 3.2.2 and
3.2.3), and run our CI tests against LibreSSL 3.2.7 (the most recent
release in the series) on Ubuntu.

LibreSSL 3.2.2 and 3.2.3 are not supported because their new X.509
certificate verifier implementation behaves sufficiently differently to
OpenSSL's (callback functions are executed twice rather than once;
nondescript errors are returned when verification fails). LibreSSL 3.2.4
reverts to the libssl-compatible verifier used in LibreSSL 3.1.

Also document the libssl-incompatible nature of the TLS 1.3
implementation in 3.1 - 3.3, and suggest falling back to a lower TLS
version if it doesn't work as expected.