GH-338 Use const qualifier with advanced callback helpers.
The three advanced callback helper functions can use the const qualifier with
their first argument 'ptr' because they do not, and never should, modify the
data 'ptr' points to.

This also helps avoiding dropping const qualifiers or using casts with functions
such as

  SSL_CTX_keylog_cb_func SSL_CTX_get_keylog_callback(const SSL_CTX *ctx);

where 'ctx' is used without const by the respective setter function

  void SSL_CTX_set_keylog_callback(SSL_CTX *ctx, SSL_CTX_keylog_cb_func cb);

In both cases 'ctx' is the value that is passed as 'ptr' to the advanced
callback helper functions.

In addition to adding 'const', the helper function comments are also fixed and
updated.