GH-352 Update documentation for library version and information functions.
Clarify documentation of OpenSSL_version_num(), SSLeay(), SSLeay_version() and
OpenSSL_version().

Add notes to OpenSSL_version_num() and SSLeay() on how to determine if the
library is OpenSSL or LibreSSL and how to interpret the version number these
functions return

With OpenSSL 3.0.0 OPENSSL_VERSION_NUMBER becomes larger than 0x20000000. For
this reason more work is needed if it's necessary to be able to tell if the
library is OpenSSL or LibreSSL.