GH-356 Work around 07_sslecho.t and 44_sess.t failures with older Perls on Windows. (#366)
Calling a one second sleep before each connect to the server seems fix the
occasionally occurring failures. Sleep is now done when running on Windows and
Perl is earlier than 5.20.0.

For the details, see GH-356 and look for CloseHandle() in Perl 5.20.0 changelog.
It seems that this is the Perl bug that causes the socket failures:
  https://github.com/Perl/perl5/issues/12979