TLS keylog test no longer requires certain TLS versions.
Test 47_keylog.t required the availability of both TLSv1.2 and TLSv1.3. TLSv1.3
has different keylog format than the older SSL and TLS versions.

The test no longer requires the both versions. If either one, or the both, are
missing, diagnostics is logged to inform that all keylog formats can not be
tested.