Silence compound-token-split-by-macro warnings with Clang >= 12
Perl's use of the GCC brace groups extension in versions prior to 5.35.2
triggers large numbers of compound-token-split-by-macro warnings
originating from perl.h when SSLeay.xs is compiled with Clang 12 or
greater. There's not much we can do about these warnings since they
originate from Perl, so before including perl.h, suppress all such
warnings if Clang >= 12 is detected.

Ideally we'd be able to limit the suppression of these warnings to
versions of Perl prior to 5.35.2, but this isn't possible due to the
PERL_VERSION_* macros being defined in perl.h.

Fixes the remainder of #383.