Remove unnecessary return value check in P_X509_get_crl_distribution_points.
P_X509_get_crl_distribution_points checked the return value of X509_get_ext_d2i
whereas P_X509_get_ext_key_usage did not. This check is no longer done because
the OpenSSL sk_*_func() functions appear to handle NULL values correctly, at
least in the case of these two P_X509_* SSLeay.xs functions.

In addition to this, the C if statement that did the NULL check had body that
was a long for loop which made the scop of the body hard to see with a quick
glance. Removing the NULL check simplifies this too.