RSA_generate_key: declare all variables in PREINIT
The default values for perl_cb and perl_data in the type signature of
RSA_generate_key cause xsubpp to emit code that sets the default values
before the CODE section. The declarations that follow in the CODE
section therefore trigger compilation failures under
-Werror=declaration-after-statement.

Move the declarations in the CODE section into the PREINIT section,
which is emitted before the default parameter values are set.

Fixes #407.