GH-394 Tests no longer depend on the legacy provider. (#423)
* GH-394 Tests no longer depend on the legacy provider.

Net::SSLeay::PEM_get_string_PrivateKey() uses DES-CBC as its default encryption
algorithm. OpenSSL 3.0 implements DES-CBC in the legacy
provider which may not be available on all systems. This commit changes
33_x509_create_cert.t PEM_get_string_PrivateKey() default encryption testing to
skip the test on systems that support providers but do not have the default
provider available.

This commit does not change the fact that PEM_get_string_PrivateKey() remains
broken on systems without the default provider when the default encryption
algorithm is used.

When obsolete encryption algorithms are updated, for example by changing
DES-CBC to AES-128-CBC, the special handling can be removed and the test can be
re-enabled on all systems.

As a reminder for this, test fail() is called when Net::SSLeay::VERSION reaches
2.x.

Closes GH-394. Relates to GH-402.

Co-authored-by: Chris Novakovic <chris@chrisn.me.uk>