Work around attempted implicit use of sha1 algorithm in test suite
The sha1 algorithm is used in key exchange for TLSv1 and TLSv1.1, so in
an environment where the sha1 algorithm is forbidden by security policy,
TLSv1 and TLSv1.1 may appear to be available in libssl but don't
actually work, with an SSL_ERROR_SSL manifesting at connect time.

This shows up in t/local/44_sess.t and t/local/45_exporter.t (though
the latter test still actually passes). These tests are updated to check
for SSL_ERROR_SSL error on client connection and skip the protocol round
if it's detected.

One additional test is added, to ensure that at least one protocol gets
the full set of tests run for it.