GH-449 Add constants for version field in certificates, certificate requests and CRLs.
The version field in certificates, certificate requests and CRLs is off-by-one
in relation to the respective constant names. For example X509_VERSION_3 has
integer value 2. Using constants is an easy way to avoid off-by-one errors.

The newly added constant are:
 - X509_VERSION_1, X509_VERSION_2 and X509_VERSION_3
 - X509_REQ_VERSION_1, X509_REQ_VERSION_2 and X509_REQ_VERSION_3
 - X509_CRL_VERSION_1 and X509_CRL_VERSION_2

These constants were added to OpenSSL 3.0.0-alpha16 and they are not available
in LibreSSL at the time of writing.