GH-452 Update threads test to avoid crash on exit.
OpenSSL 3.0 added atexit() handler that appears to cause crashes with test
62_threads-ctx_new-deadlock.t.

The test is now updated to initialise OpenSSL libcrypto so that atexit()
handler is not called. As a result, OpenSSL libraries are no longer cleaned up
on exit. This is fine for the test since the process that runs it is already
exiting and the test requires no special cleaning up.