GH-454 Tone down warning against mixing compilers and compiler flags.
With Linux and macOS, for example, compiling OpenSSL and Net::SSLeay with gcc
and clang appears to work these days. It's even possible to compile Net::SSLeay
with a C++ compiler and link it against OpenSSL compiled with a C compiler.

On some other platforms it may still be required to use matching compiler and
flags, but it seems not to be universally required anymore.