Don't automatically detect libssl/libcrypto with ExtUtils::PkgConfig
This reverts da6292e138644aa1cdd234114c6198675611693c.

The automatic detection of libssl/libcrypto with pkg-config conflicts
with the OPENSSL_PREFIX method of specifying where Makefile.PL should
look for libssl/libcrypto. The one known failure case at the moment is
when OPENSSL_PREFIX points to a copy of OpenSSL that has compression
enabled, but pkg-config points to a copy of OpenSSL that doesn't have
compression enabled - this causes "-lz" to be omitted from the list of
options passed to the linker, causing a compilation failure. There may
be other failure cases.

The automatic detection of libssl/libcrypto is a very useful feature, so
we'll find and fix all the edge cases and bring it back in time for
Net-SSLeay 1.96.

Fixes #465.