Replace binary test file with ASCII text file
t/data/binary-test.file is a binary blob of unknown provenance. To
provide more reassurance that nothing malicious is hidden in binary
blobs in the test suite, replace this file with the ASCII text file
t/data/alphanum.txt, which consists of repetitions of the English
alphabet and the Arabic numerals with lines wrapped at 80 characters.

Closes #475.