GH-521 Increase allowed timeout in 62_threads-ctx_new-deadlock.t.
Test 62_threads-ctx_new-deadlock.t times out on slow systems such as GitHub
Actions ubuntu-latest (24.04) runner with QEMU and Docker to run Ubuntu 24.04
on ARMv64/v8.

Each test seems to take several seconds and the current default timeout of 20
seconds is not long enought to allow the test to finish. Increasing it to 30
seconds seems to make the problem go away.